+++ src/main/java/com/farcr/nomansland/datagen/sound/NMLSoundProvider.java
@@ -190,18 +190,28 @@ public class NMLSoundProvider extends LodestoneBlockSoundEventSystem {
         ));
 
         add(NMLSounds.CAVE_DEPTH_MUSIC, s -> s.subtitle(null).with(
-                sound("minecraft:music/game/oxygene").stream(),
-                sound("minecraft:music/game/key").stream(),
-                sound("minecraft:music/game/minecraft").stream(),
-                sound("minecraft:music/game/mice_on_venus").stream(),
-                sound("minecraft:music/game/dry_hands").stream(),
-                sound("minecraft:music/game/yakusoku").stream().volume(0.4f),
-                sound("minecraft:music/game/endless").stream().volume(0.4f)
+                sound("minecraft:music/game/oxygene").weight(5).stream(),
+                sound("minecraft:music/game/key").weight(5).stream(),
+                sound("minecraft:music/game/minecraft").weight(3).stream(),
+                sound("minecraft:music/game/mice_on_venus").weight(5).stream(),
+                sound("minecraft:music/game/dry_hands").weight(5).stream(),
+                sound("minecraft:music/game/yakusoku").weight(3).stream().volume(0.4f),
+                sound("minecraft:music/game/endless").weight(3).stream().volume(0.4f),
+                sound("nomansland:music/tarnished").weight(1).stream()
+        ));
+
+        add(NMLSounds.ALCHEMIST_RUINS_MUSIC, s -> s.subtitle(null).with(
+                sound("minecraft:music/game/oxygene").weight(5).stream(),
+                sound("minecraft:music/game/key").weight(5).stream(),
+                sound("minecraft:music/game/minecraft").weight(3).stream(),
+                sound("minecraft:music/game/mice_on_venus").weight(3).stream(),
+                sound("minecraft:music/game/dry_hands").weight(3).stream(),
+                sound("nomansland:music/tarnished").weight(20).stream()
         ));
 
         add(NMLSounds.MUSIC_DISC_GUIDANCE, s -> s.subtitle(null).with(sound("records/guidance").stream()));
 
-        add(NMLSounds.SHROOMANIAC_TARNISHED, s -> s.subtitle(null).with(sound("music/shroomaniac_tarnished").volume(0.1f).stream()));
+        add(NMLSounds.MUSIC_TARNISHED, s -> s.subtitle(null).with(sound("music/tarnished").volume(0.1f).stream()));
 
         //Misc
         add(NMLSounds.DROPLET_FALLS, s -> s.with(allSounds("minecraft:block/beehive/drip", se -> se.attenuationDistance(8).volume(0.3f))));