+++ changelog.md
@@ -2,4 +2,5 @@
 - Fixed occasional client crash with inverted bells
 - Made it possible to re-trigger the dream
 - Updated to Mixed Litter 1.2.1
-- Added the sun painting to the list of survival-obtainable paintings
+- Added "Sun" and "White Eyes" paintings to the list of survival-obtainable paintings
+- Changed Alchemist Ruins to use their own music pool, where "Tarnished" by Shroomaniac will play most often